rewrite!: Refactor core of the library
The main change here is how we construct and create the template. This
rewrite is to make the process involved more involved instead of having
two different modes. Also, strings are now escaped by default.

Diffing is also rewrote to include text nodes. This means that a way to
differentiate must be created. We do this with the key property in __meta.

One objective for this rewrite is to make future development easier and
to allow flexibility for a possible plugin system.

Closes #29 #30 #34