revert: "feat: Add if-else directive"
Roll back this feature because it is too premature. The cases it can be
used is very limited. At the moment, we can only use this for static elements
and also, we need the directive to be present in the markup right from
the start which is very restraining. We also need to figure out how to
seamlessly integrate this custom-components.

This will live on a separate branch for now till ready to merge.

This reverts commit 2ced458d4e9303d9e5a9b49497c8b133fa5b2df1.