refactor(diffing): Slightly change implementation
We relegate responsibility for any changes to dynamic children to hooks
so they are now ignored in diffing. This also means that for first
rearrange and patch, only the dynamic nodes are actually involved (see
see changes in main.js and modify.js).

This commit also hopefully fixes the bugs left